feat: 추천 상품 담기 기능 추가

+++ app/src/main/java/woowacourse/shopping/data/datasource/CartItemDataSource.kt
@@ -13,7 +13,7 @@ interface CartItemDataSource {
 
     fun submitCartItem(
         cartItem: CartItemRequest,
-        callback: () -> Unit,
+        callback: (Long) -> Unit,
     )
 
     fun removeCartItem(
